src/paskoocheh/requirements.txt
amqp==5.1.1
attrs==19.3.0
beautifulsoup4==4.8.1
billiard==3.6.1.0
bleach==3.3.0
boto3==1.26.47
botocore==1.29.47
certifi==2023.07.22
cffi==1.13.2
chardet==3.0.4
cryptography==41.0.7
Django==3.2.23
django-cors-headers==4.0.0
django-debug-toolbar==3.8.1
django-debug-toolbar-template-profiler==2.1.0
django-debug-toolbar-template-timings==0.9
django-fancy-cache==1.2.1
django-hosts==5.1
django-jalali==5.0.0
django-markdownx==3.0.1
django-pipeline==2.1.0
django-redis==5.0.0
django-sslserver==0.21
django-storages==1.13.2
djangorestframework==3.12.4
docutils==0.15.2
gunicorn==20.1.0
idna==2.8
importlib-metadata==0.23
jdatetime==3.7.0
jmespath==0.9.4
kombu==5.2.1
Markdown==3.3
markdownify==0.4.1
more-itertools==7.2.0
PGPy==0.6.0
psycopg2==2.8.4
pyasn1==0.4.8
pycparser==2.19
PyICU==2.11
python-dateutil==2.8.0
pytz==2019.3
redis==4.4.4
requests==2.31.0
s3transfer==0.6.0
six==1.16.0
soupsieve==1.9.5
sqlparse==0.4.4
ua-parser==0.8.0
urllib3==1.25.11
vine==5.0.0
virtualenv==16.7.8
webencodings==0.5.1
wrapt==1.11.2
zipp==0.6.0
html2text==2020.1.16
telegraph==1.4.0
pyaxmlparser==0.3.24
protobuf==3.18.3
Pygments==2.15.1
python-magic==0.4.27
typer==0.9.0
rich==13.4.2
strawberry-graphql==0.219.2
strawberry-django-auth==0.375.0
strawberry-graphql-django==0.31.0
wagtail==4.1.9
wagtail-markdown==0.11.0
